Port to Python>=3.4 and make more Pythonic/maintainable
This commit includes all of the changes I have made to the Metrics tool
over the past few months. Everything was modified in some way to make
the overall project more maintainable and conform to standard Python 3
conventions. With the old package being stuck with Python 2, I have
updated it so that it now functions under Python 3.4 and greater. All in
all, I believe the quality of the code has increased greatly and it is
much more managable. Finally, the original tests are back and even more
have been added which all work under recent versions of Pytest.